Move shared logic into AbstractReconcileTransaction to avoid mistakes
